Mailinglist header should not be editable for recipients (#5299)
* Reduce mailing list header

* Rename MessageListProfile file

* Remove unused hook
